requirements.txt
requests
google.generativeai
streamlit

